requirements.txt
spotipy
numpy
opencv-python
keras
tensorflow
flask
PyExecJS
python_dotenv